fix(inkling): drop redundant i32 cast in audio test helper
clippy --all-features -D warnings rejects the no-op i32->i32 cast in
seeded_signal; the value is already i32.

+++ crates/multimodal/src/audio/processors/inkling.rs
@@ -369,7 +369,7 @@ mod tests {
     fn seeded_signal(scale: f32) -> Vec<f32> {
         (0..1600)
             .map(|i| {
-                let raw = ((i * 73 + 17 * 977) % 65_536) as i32 - 32_768;
+                let raw = (i * 73 + 17 * 977) % 65_536 - 32_768;
                 raw as f32 / 32_768.0 * scale
             })
             .collect()
